+++ .github/CHANGELOG.md
@@ -112,6 +112,7 @@ Always start an item with "1."; GitHub will number the rendered list automatical
 1. [ND] Smooth ILS and VOR deviations on ND - @BlueberryKing (BlueberryKing)
 1. [A32NX/MCDU] Fix navigation arrows missing from remote MCDU pages - @tracernz (Mike)
 1. [A32NX/FWS] Add EWD overflow logic - @FozzieHi (fozzie)
+1. [EFB] Added fix to make sure all doors are closed when pushing back - @frankkopp (cdr_maverick)
 
 ## 2024.1.0